Adds ability to supply source parameters to embed code
This commit adds the IframeBuilder class, which can distinguish between iframe
attributes and a provided list of whitelisted parameters  that can optionally be
appended to the source of the iframe. This provides a wrapper around services'
JavaScript APIs.

This commit also provides documentation for the intended workings of this
feature - though it does not fully flesh out the functionality.